Better keys/certs directory structure

+++ certs/gen_node.sh
@@ -11,36 +11,42 @@ if [ -z "$CLUSTER" ] || [ -z "$NODE" ]; then
 	exit 1
 fi
 
+if [ ! -d "$CLUSTER" ]; then
+	echo "Please create cluster first"
+	exit 1
+fi
+
 rm -f \
-	"${NODE}-key.pem" "${NODE}-cert.pem" \
-	"${NODE}.pkcs12" "${NODE}-keystore.jks"
+	"${CLUSTER}/${NODE}-key.pem" "${CLUSTER}/${NODE}-cert.pem" \
+	"${CLUSTER}/${NODE}.pkcs12" "${CLUSTER}/${NODE}-keystore.jks"
 
 echo "* Generate node key and certificate request"
 openssl req \
 	-newkey rsa:4096 -subj "/O=${CLUSTER}/OU=${NODE}" -nodes \
-	-keyout "${NODE}-key.pem" -out "${NODE}-req.pem"
+	-keyout "${CLUSTER}/${NODE}-key.pem" -out "${CLUSTER}/${NODE}-req.pem"
 
 echo "* Sign node certificate request with cluster key-certificate"
 openssl x509 \
-	-CA "${CLUSTER}-cert.pem" -CAkey "${CLUSTER}-key.pem" -CAcreateserial \
-	-req -in "${NODE}-req.pem" -out "${NODE}-cert.pem" -days "${VALIDITY}"
-rm "${NODE}-req.pem"
+	-CA "${CLUSTER}/cluster-cert.pem" -CAkey "${CLUSTER}/cluster-key.pem" -CAcreateserial \
+	-req -in "${CLUSTER}/${NODE}-req.pem" -out "${CLUSTER}/${NODE}-cert.pem" -days "${VALIDITY}"
+rm "${CLUSTER}/${NODE}-req.pem"
 
 echo "* Package node PKCS12 key-certificate"
 openssl pkcs12 -export \
-	-name "${NODE}" -in "${NODE}-cert.pem" -inkey "${NODE}-key.pem" \
-	-out "${NODE}.pkcs12" -passout pass:
+	-in "${CLUSTER}/${NODE}-cert.pem" -inkey "${CLUSTER}/${NODE}-key.pem" \
+	-name "${NODE}" -out "${CLUSTER}/${NODE}.pkcs12" -passout pass:""
 
 echo "* Add node PKCS12 key-certificate to node Java keystore"
-keytool -importkeystore -alias "${NODE}" \
-	-srckeystore "${NODE}.pkcs12" -srcstorepass "" -srcstoretype pkcs12 \
-	-destkeystore "${NODE}-keystore.jks" -deststorepass "${PASSWORD}"
+keytool \
+	-importkeystore -alias "${NODE}" \
+	-srckeystore "${CLUSTER}/${NODE}.pkcs12" -srcstorepass "" -srcstoretype pkcs12 \
+	-destkeystore "${CLUSTER}/${NODE}-keystore.jks" -deststorepass "${PASSWORD}"
 
 echo "* Add cluster certificate to node Java keystore"
 keytool \
-	-importcert -alias "${CLUSTER}" -file "${CLUSTER}-cert.pem" -noprompt \
-	-keystore "${NODE}-keystore.jks" -storepass "${PASSWORD}"
+	-importcert -alias "${CLUSTER}" -file "${CLUSTER}/cluster-cert.pem" -noprompt \
+	-keystore "${CLUSTER}/${NODE}-keystore.jks" -storepass "${PASSWORD}"
 
 echo "* Display node Java keystore"
 keytool -list \
-	-keystore "${NODE}-keystore.jks" -storepass "${PASSWORD}"
+	-keystore "${CLUSTER}/${NODE}-keystore.jks" -storepass "${PASSWORD}"
